agents/profiler/requirements.txt
boto3>=1.34
dnspython>=2.6
httpx>=0.27
psycopg[binary,pool]>=3.1
pydantic>=2.7
pydantic-settings>=2.3
python-jose[cryptography]>=3.3
structlog>=24.1
tenacity>=8.2
